Fix warnings

+++ src/Collapse/utilities/distance_matrix_edge_collapse_rips_persistence.cpp
@@ -93,7 +93,7 @@ int main(int argc, char* argv[]) {
   Flag_complex_sparse_matrix flag_complex(proximity_graph);
 
   Simplex_tree stree;
-  for (Vertex_handle vertex = 0; vertex < distances.size(); vertex++) {
+  for (Vertex_handle vertex = 0; static_cast<std::size_t>(vertex) < distances.size(); vertex++) {
     // insert the vertex with a 0. filtration value just like a Rips
     stree.insert_simplex({vertex}, 0.);
   }
