+++ src/Collapse/utilities/point_cloud_edge_collapse_rips_persistence.cpp
@@ -80,7 +80,7 @@ int main(int argc, char* argv[]) {
   Flag_complex_sparse_matrix mat_filt_edge_coll(proximity_graph);
 
   Simplex_tree stree;
-  for (Vertex_handle vertex = 0; vertex < point_vector.size(); vertex++) {
+  for (Vertex_handle vertex = 0; static_cast<std::size_t>(vertex) < point_vector.size(); vertex++) {
     // insert the vertex with a 0. filtration value just like a Rips
     stree.insert_simplex({vertex}, 0.);
   }
